Split up QuerySelector into new FieldReplacementProcessor
