+++ src/Firestorm.Engine.Implementation/Firestorm.Engine.Implementation.csproj
@@ -1,13 +1,10 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;net452</TargetFrameworks>
     <RootNamespace>Firestorm.Engine</RootNamespace>
   </PropertyGroup>
-
   <ItemGroup>
     <ProjectReference Include="..\Firestorm.Core\Firestorm.Core.csproj" />
     <ProjectReference Include="..\Firestorm.Engine\Firestorm.Engine.csproj" />
   </ItemGroup>
-
 </Project>
